refactor: reorganizes code
